+++ packages/backend/src/nest/registration/registration.service.ts
@@ -32,9 +32,11 @@ export class RegistrationService extends EventEmitter implements OnModuleInit {
     }
     const pendingCsrs = await extractPendingCsrs(payload)
 
-    pendingCsrs.forEach(async csr => {
-      await this.registerUserCertificate(csr)
-    })
+    await Promise.all(
+      pendingCsrs.map(async csr => {
+        await this.registerUserCertificate(csr)
+      })
+    )
 
     if (payload.id) this.emit(RegistrationEvents.FINISHED_ISSUING_CERTIFICATES_FOR_ID, { id: payload.id })
   }