+++ packages/desktop/src/renderer/components/widgets/possibleImpersonationAttackModal/PossibleImpersonationAttackModal.component.tsx
@@ -40,14 +40,12 @@ const StyledGrid = styled(Grid)(({ theme }) => ({
 
 export interface PossibleImpersonationAttackModalComponentProps {
   communityName: string
-  leaveCommunity: () => void
   open: boolean
   handleClose: () => void
 }
 
 const PossibleImpersonationAttackModalComponent: React.FC<PossibleImpersonationAttackModalComponentProps> = ({
   communityName,
-  leaveCommunity,
   handleClose,
   open,
 }) => {
@@ -69,9 +67,7 @@ const PossibleImpersonationAttackModalComponent: React.FC<PossibleImpersonationA
           <br />
           <strong>This should never happen and we recommend leaving this community immediately!</strong>
         </Typography>
-        <Button className={classes.button} data-testid='unregistered-button' onClick={leaveCommunity}>
-          Leave community
-        </Button>
+        {/* Temporarily hiding button - https://github.com/TryQuiet/quiet/issues/2025 */}
       </StyledGrid>
     </Modal>
   )